Story1,9,10 update

+++ app/src/main/java/com/example/wwr/Route.java
@@ -42,7 +42,7 @@ public class Route {
         this.totalSeconds = seconds;
     }
 
-    public void updateMiles(long miles) {
+    public void updateMiles(double miles) {
         this.totalMiles = miles;
     }
